Cache aspects plus diable pretty XML by default.
Aspects loaded from a file are cached in a thread-safe datastructure, removing
the need of re-parsing this file every time. For long running processes (e.g.
using opener-daemons) this should save quite the disk activity.

Another change is the disabling of pretty XML by default. REXML is really slow
at formatting but we still want somewhat sane XML output for the tests. As a
compromise the pretty printing has been disabled by default and is enabled only
in the tests and the CLI.

Last, I've added a simple benchmark-ips based benchmark that we can use to track
the performance between changes.